Updated various files and cleaned up the codebase
